requirements.txt
ImageHash
numpy
opencv-python
pillow
pytesseract
fastapi
uvicorn
python-multipart
requests